feat 知识库上传接口完成
- 上传之后自动嵌入指定的向量数据库中

+++ tests/test_kb_router.py
@@ -5,18 +5,20 @@ from fastapi.testclient import TestClient
 
 from common import ResponseCode
 from kb.kb_config import UploadConfig
-from kb.kb_router import router
+from kb.kb_router import router, write_to_kb_with_docx
 
 client = TestClient(router)
 
 
 class Test(TestCase):
+    test_file = 'test.docx'
+
     def test_upload_file(self):
         """测试文件上传的正常响应"""
-        docx_file_path = 'test.docx'
+        docx_file_path = Test.test_file
         with open(docx_file_path, 'rb') as docx_file:
             docx_content = docx_file.read()
-        resp = client.post('/kb/upload/',
+        resp = client.post('/kb/upload/Hello;bge-m3',
                            files={"file": ('example.docx', docx_content,
                                            "application/vnd.openxmlformats-officedocument.wordprocessingml.document")})
         self.assertEqual(resp.status_code, 200)
@@ -26,4 +28,10 @@ class Test(TestCase):
         os.remove(file_path)
 
     def tearDown(self):
-        os.rmdir(UploadConfig.UPLOAD_SAVING_PATH)
+        if os.path.exists(UploadConfig.UPLOAD_SAVING_PATH):
+            os.rmdir(UploadConfig.UPLOAD_SAVING_PATH)
+
+    def test_write_to_kb_with_docx(self):
+        docx_file_path = Test.test_file
+        kb_id = write_to_kb_with_docx(docx_file_path, "Hello;bge-m3")
+        self.assertEqual("Hello;bge-m3", kb_id)
